add putSubmission function for updating submission data
This commit adds a new function, putSubmission, to the API module. The function is responsible for making a PUT request to update the submission data on the server. It takes in parameters such as broker_submission_id, target, embargo, and data. The function uses axios to send the request with the necessary headers and authorization token. The response data is then returned.